position added

+++ map/StripTable.h
@@ -22,9 +22,9 @@ public:
 	float interX(int sn);
 	float interY(int sn, int snPerp);
 
-	std::vector<double> pos(int sn, char axis);
-	std::vector<double> posX(int sn){return this->pos(sn, 'x');};
-	std::vector<double> posY(int sn){return this->pos(sn, 'y');};
+	std::vector<double> pos(double sn, char axis);
+	std::vector<double> posX(double sn){return this->pos(sn, 'x');};
+	std::vector<double> posY(double sn){return this->pos(sn, 'y');};
 
 	int zone(int snx, int sny);
 	int nbZone(){return zonePitch.size();};
